requirements.txt
blinker==1.8.2
click==8.1.7
comtypes==1.4.4
Flask==3.0.3
gunicorn==22.0.0
imageio==2.34.1
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
MouseInfo==0.1.3
mss==9.0.1
numpy==2.0.0
packaging==24.1
pillow==10.3.0
psutil==6.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.2.0
pywinauto==0.6.6
rubicon-objc==0.4.9
setuptools==70.0.0
six==1.16.0
Werkzeug==3.0.3
